Reformat code to fit Kotlin style.

+++ kubed-core/src/main/kotlin/kubed/interpolate/color/Rgb.kt
@@ -9,9 +9,9 @@ fun interpolateRgb(start: Any, end: Any, gamma: Double = 1.0): (Double) -> Color
     val sc = Rgb.convert(start)
     val ec = Rgb.convert(end)
 
-    val r = gamma(gamma)(sc.r.toDouble(), ec.r.toDouble())
-    val g = gamma(gamma)(sc.g.toDouble(), ec.g.toDouble())
-    val b = gamma(gamma)(sc.b.toDouble(), ec.b.toDouble())
+    val r = gamma(gamma)(sc.r, ec.r)
+    val g = gamma(gamma)(sc.g, ec.g)
+    val b = gamma(gamma)(sc.b, ec.b)
     val opacity = nogamma(sc.opacity, ec.opacity)
 
     return { t -> Color(r(t), g(t), b(t), opacity(t)) }
@@ -28,9 +28,9 @@ private fun interpolate(colors: List<Any>, spline: (DoubleArray) -> (Double) -> 
 
     for(i in colors.indices) {
         val rgb = Rgb.convert(colors[i])
-        r[i] = rgb.r.toDouble()
-        g[i] = rgb.g.toDouble()
-        b[i] = rgb.b.toDouble()
+        r[i] = rgb.r
+        g[i] = rgb.g
+        b[i] = rgb.b
     }
 
     val rs = spline(r)
